Create README.md
Objective of this exercise is to learn continuous build process (CB) and continuous integration (CI).
You are expected to:
- Write python program to convert Arabic Numerals to Roman Numerals
-  Integrate Continuous Build Process to check if your software in each development stage passed the build process.
-  Integrate unit test and run the unit test in continuous integration process.
Deadline for completion of this project is January 29th, 2019.
You are expected to show all aspects of your work.  This includes results of build process.
You are expected to use Github actively during this exercise.